re-wrote to hooks
