+++ src/constants.js
@@ -82,28 +82,3 @@ export function passProps(props) {
     return passedCopy;
   }, {});
 }
-
-// Polyfill
-if (!Array.prototype.find) {
-  /* eslint no-extend-native: [0], prefer-rest-params: [0] */
-  Array.prototype.find = function find(predicate) {
-    if (this == null) {
-      throw new TypeError('Array.prototype.find called on null or undefined');
-    }
-    if (typeof predicate !== 'function') {
-      throw new TypeError('predicate must be a function');
-    }
-    const list = Object(this);
-    const length = list.length >>> 0; // eslint-disable-line
-    const thisArg = arguments[1];
-    let value;
-
-    for (let i = 0; i < length; i++) {
-      value = list[i];
-      if (predicate.call(thisArg, value, i, list)) {
-        return value;
-      }
-    }
-    return undefined;
-  };
-}
